Start downloading data from backend when loading splash screen
