+++ app/src/main/java/com/example/quiz/categorylist/CategoryListFragment.kt
@@ -29,7 +29,6 @@ class CategoryListFragment : BaseFragment<CategoryListViewModel, CategoryListFra
         with(category_list) {
             adapter = CategoryListAdapter().apply {
                 viewModel.categoryList.observe(this@CategoryListFragment, Observer {
-                    viewModel.downloadAll()
                     submitList(it)
                 })
             }
